requirement.txt
kafka-python==2.0.2; python_version < '3.12'
kafka-python @ git+https://github.com/dpkp/kafka-python.git ; python_version >= '3.12'
yfinance
mysql-connector-python
pandas
pymongo